Correct initial chat history loading
Server now do not send room list by default, only on request
Reducer is modified. Now room is changed only on it's action.
Middleware modified. When user is logged in - request rooms,
When rooms are received - set first one as current via action

So, when users logs on it receives room list. When room list is received
first items is becomes current. When room is selected - history is loaded